refactor: extract FunctionRegistryTrie and BoundedASTCache into own modules
